1. 将付费改为0.01分 2. 使用druid(主要是要监控) 3. 删除一些不必要的拆包

+++ behappy-hosp/src/main/java/org/xiaowu/behappy/hosp/config/HospitalReceiver.java
@@ -55,7 +55,7 @@ public class HospitalReceiver {
                 schedule.setAvailableNumber(orderMqVo.getAvailableNumber());
             } else {
                 //取消预约更新预约数
-                int availableNumber = schedule.getAvailableNumber().intValue() + 1;
+                int availableNumber = schedule.getAvailableNumber() + 1;
                 schedule.setAvailableNumber(availableNumber);
             }
             scheduleService.update(schedule);
